requirements.txt
fastapi
uvicorn
python-multipart
jinja2
python-dotenv
openai>=1.0.0